Large Portion of Implementation Completed
Another big commit. This one consists of the final xml formats for the
Kobayashi manufacturer xmls as well as the warehouse items xml. Finished
implementing a large portion of the code. Only 2 methods remain to be
completed within the KobayashiManufacturerService. Once these are
completed, testing will need to be done on all the classes.